userspace/nat: split source-NAT compiler into nat_source.go
Pure code-motion refactor of the multi-family userspace NAT snapshot
compiler (#4421, codex-173 #5). nat.go had grown to ~1286 lines spanning
source/static/destination/NAT64/NPTv6 snapshot builders plus their
type-local helpers. This first increment moves the source-NAT
declarations verbatim into a sibling nat_source.go in the same package:

  buildSourceNATSnapshots, buildSourceNATSnapshotsWithFeeds, the
  snatTier* scope-specificity consts, sourceNATScopeTier,
  scopeContextTier, the natProtoAny/natProtoNever match sentinels,
  sourceNATDestPortRanges, natAppProtoNumber, buildSourceNATAppTerms,
  and sourceNATPoolPortRange.

No declaration body is edited and nothing is renamed. Cross-type shared
helpers (natCounterID, the address-name resolvers, coalescePortRanges,
appPortsFromSpec, natNeverMatchPortRange) stay in the residual nat.go so
they are not duplicated or stranded. Imports were pruned per file with
goimports; the sorted top-level production decl-name set is unchanged.

Validation: go build ./... green.
